Initial work done on including also a coin exchange provider
