+++ apps/migrator/sql/V1__Create_basic_tables.sql
@@ -511,6 +511,18 @@ CREATE TRIGGER notify_token_exchange_rates
   AFTER INSERT OR UPDATE OR DELETE ON token_exchange_rates
   FOR EACH ROW EXECUTE PROCEDURE notify_event();
 
+
+/* Coin exchange rates table */
+CREATE TABLE coin_exchange_rates
+(
+  id           CHAR(6) PRIMARY KEY,
+  rate         NUMERIC NOT NULL,
+  market_cap   NUMERIC NOT NULL,
+  vol24h       NUMERIC NOT NULL,
+  change24h    NUMERIC NOT NULL,
+  last_updated BIGINT  NOT NULL
+);
+
 /* metrics hyper tables */
 CREATE TABLE block_metrics_header
 (
