+++ apps/processing/connectors/sources/exchanges/src/test/kotlin/com/ethvm/kafka/connect/sources/exchanges/provider/CoinGeckoCurrencyExchangeProviderTest.kt
@@ -19,26 +19,26 @@ import org.apache.kafka.connect.errors.RetriableException
 import org.apache.kafka.connect.source.SourceRecord
 import java.io.IOException
 
-class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
+class CoinGeckoCurrencyExchangeProviderTest : BehaviorSpec() {
 
   init {
 
-    Given("an empty CoinGeckoTokenExchangeProvider") {
+    Given("an empty CoinGeckoCurrencyExchangeProvider") {
 
-      val provider = CoinGeckoTokenExchangeProvider()
+      val provider = CoinGeckoCurrencyExchangeProvider()
 
-      When("we fetch for token exchange rates") {
+      When("we fetch for coin exchange rates") {
 
         val records: List<SourceRecord> = provider.fetch()
 
-        Then("we should parse default token exchanges size") {
+        Then("we should parse default coin exchanges") {
           records shouldNotBe null
-          records.size shouldBe 732
+          records.size shouldBe 3
         }
       }
     }
 
-    Given("a configured CoinGeckoTokenExchangeProvider but without tokens ids") {
+    Given("a configured CoinGeckoCurrencyExchangeProvider but without coins ids") {
 
       val inputStreamProvider = InputStreamProvider { path -> javaClass.getResourceAsStream("/$path") }
 
@@ -46,29 +46,28 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
         .addInterceptor(OkHttpMockInterceptor(inputStreamProvider, 0))
         .build()
 
-      val provider = CoinGeckoTokenExchangeProvider(
+      val provider = CoinGeckoCurrencyExchangeProvider(
         mapOf(
-          "topic" to "token-exchange-rates",
+          "topic" to "coin-exchange-rates",
           "currency" to "usd",
-          "perPage" to 10,
-          "tokenIds" to emptyList<TokenIdEntry>()
+          "coinIds" to emptyList<TokenIdEntry>()
         ),
         okHttpClient,
         CoinGeckoExchangeProvider.jackson
       )
 
-      When("we fetch for token exchange rates") {
+      When("we fetch for coin exchange rates") {
 
         val records: List<SourceRecord> = provider.fetch()
 
-        Then("we should parse empty token exchanges") {
+        Then("we should parse empty coin exchanges") {
           records shouldNotBe null
           records.size shouldBe 0
         }
       }
     }
 
-    Given("a properly configured CoinGeckoTokenExchangeProvider with tokens ids") {
+    Given("a properly configured CoinGeckoCurrencyExchangeProvider with coins ids") {
 
       val inputStreamProvider = InputStreamProvider { path -> javaClass.getResourceAsStream("/$path") }
 
@@ -77,36 +76,24 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
         .build()
 
       val tokensIds = listOf(
-        TokenIdEntry("hurify", "0xcdb7ecfd3403eef3882c65b761ef9b5054890a47"),
-        TokenIdEntry("iqeon", "0x0db8d8b76bc361bacbb72e2c491e06085a97ab31"),
-        TokenIdEntry("1sg", "0x0f72714b35a366285df85886a2ee174601292a17"),
-        TokenIdEntry("oxbitcoin", "0xB6eD7644C69416d67B522e20bC294A9a9B405B31"),
-        TokenIdEntry("first-blood", "0xAf30D2a7E90d7DC361c8C4585e9BB7D2F6f15bc7"),
-        TokenIdEntry("parkgene", "0x6dd4e4aad29a40edd6a409b9c1625186c9855b4d"),
-        TokenIdEntry("1world", "0xfdbc1adc26f0f8f8606a5d63b7d3a3cd21c22b23"),
-        TokenIdEntry("accelerator-network", "0x13f1b7fdfbe1fc66676d56483e21b1ecb40b58e2"),
-        TokenIdEntry("arcblock", "0xb98d4c97425d9908e66e53a6fdf673acca0be986"),
-        TokenIdEntry("the-abyss", "0x0e8d6b471e332f140e7d9dbb99e5e3822f728da6"),
-        TokenIdEntry("adelphoi", "0x660e71483785f66133548b10f6926dc332b06e61"),
-        TokenIdEntry("adbank", "0x2baac9330cf9ac479d819195794d79ad0c7616e3")
+        TokenIdEntry("hurify", "0xcdb7ecfd3403eef3882c65b761ef9b5054890a47")
       )
 
-      val provider = CoinGeckoTokenExchangeProvider(
+      val provider = CoinGeckoCurrencyExchangeProvider(
         mapOf(
-          "topic" to "token-exchange-rates",
+          "topic" to "coin-exchange-rates",
           "currency" to "usd",
-          "perPage" to 10,
-          "tokenIds" to tokensIds
+          "coinIds" to tokensIds
         ),
         okHttpClient,
         CoinGeckoExchangeProvider.jackson
       )
 
-      When("we fetch for token exchange rates") {
+      When("we fetch for coin exchange rates") {
 
         val records: List<SourceRecord> = provider.fetch()
 
-        Then("we should parse correctly token exchanges rates") {
+        Then("we should parse correctly coin exchanges rates") {
           records shouldNotBe null
           records.size shouldBe 12
 
@@ -125,7 +112,7 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
       }
     }
 
-    Given("a properly second configured CoinGeckoTokenExchangeProvider with tokens ids") {
+    Given("a properly second configured CoinGeckoCurrencyExchangeProvider with tokens ids") {
 
       val okHttpClient = mockk<OkHttpClient>()
 
@@ -146,11 +133,10 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
         TokenIdEntry("hurify", "0xcdb7ecfd3403eef3882c65b761ef9b5054890a47")
       )
 
-      val provider = CoinGeckoTokenExchangeProvider(
+      val provider = CoinGeckoCurrencyExchangeProvider(
         mapOf(
           "topic" to "token-exchange-rates",
           "currency" to "usd",
-          "perPage" to 10,
           "tokenIds" to tokensIds
         ),
         okHttpClient,
@@ -167,7 +153,7 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
       }
     }
 
-    Given("a properly third configured CoinGeckoTokenExchangeProvider with tokens ids") {
+    Given("a properly third configured CoinGeckoCurrencyExchangeProvider with coins ids") {
 
       val okHttpClient = mockk<OkHttpClient>()
 
@@ -188,12 +174,11 @@ class CoinGeckoTokenExchangeProviderTest : BehaviorSpec() {
         TokenIdEntry("hurify", "0xcdb7ecfd3403eef3882c65b761ef9b5054890a47")
       )
 
-      val provider = CoinGeckoTokenExchangeProvider(
+      val provider = CoinGeckoCurrencyExchangeProvider(
         mapOf(
-          "topic" to "token-exchange-rates",
+          "topic" to "coin-exchange-rates",
           "currency" to "usd",
-          "perPage" to 10,
-          "tokenIds" to tokensIds
+          "coinIds" to tokensIds
         ),
         okHttpClient,
         CoinGeckoExchangeProvider.jackson
