+++ apps/processing/connectors/sources/exchanges/src/test/kotlin/com/ethvm/kafka/connect/sources/exchanges/provider/CoinGeckoExchangeProviderTest.kt
@@ -1,0 +1,6 @@
+package com.ethvm.kafka.connect.sources.exchanges.provider
+
+import io.kotlintest.specs.BehaviorSpec
+
+class CoinGeckoExchangeProviderTest : BehaviorSpec() {
+}
